renamed files and functions and classes and ids to reflect change from CriminalOrWitness toggle to ActiveList toggle
