+++ index.html
@@ -15,7 +15,7 @@
     <main class="container">
         <div class="content--left">
             <div class="layoutControls">
-                <div class="toggleCriminalsOrWitnessesButtonContainer"></div>
+                <div class="toggleActiveListButtonContainer"></div>
                 <div class="toggleNotesDisplayButtonContainer"></div>
             </div>
             <div class="filters">
