+++ scripts/layout/layoutChangeHandler.js
@@ -24,9 +24,9 @@ eventHub.addEventListener('notesToggled', event => {
 });
 
 /**
- * Listen for the criminals or witness list toggle event - render the proper list and hide/show the filters depending on which list chosen
+ * Listen for the list change event - swap out which list is currently being rendered in the main app area
  */
-eventHub.addEventListener('criminalsOrWitnessesToggled', event => {
+eventHub.addEventListener('listChanged', event => {
   const listToDisplay = event.detail.listToDisplay;
 
   if(listToDisplay === 'criminalList') {
