+++ scripts/main.js
@@ -2,7 +2,7 @@ import { ConvictionSelect } from './convictions/ConvictionSelect.js';
 import { OfficerSelect } from './officers/OfficerSelect.js';
 import { CriminalList } from './criminals/CriminalList.js';
 import { ToggleNotesDisplayButton } from './layout/ToggleNotesDisplayButton.js';
-import { ToggleCriminalsOrWitnessesButton } from './layout/ToggleCriminalsOrWitnessesButton.js';
+import { ToggleActiveListButton } from './layout/ToggleActiveListButton.js';
 import './notes/NoteForm.js';
 import './notes/NoteList.js';
 import './criminals/AssociateList.js';
@@ -12,4 +12,4 @@ ConvictionSelect();
 OfficerSelect();
 CriminalList();
 ToggleNotesDisplayButton();
-ToggleCriminalsOrWitnessesButton();
+ToggleActiveListButton();
